Fix a couple of issues with picking on glyphed points.
Update VTK reader to set mesh type to AVT_POINT_MESH when data is poly data and contains only vertex cells.
Added GetBounds override to vtkPointGlyphMapper so it can return the bounds of the underlying mapper used.
Updated GlyphPick in VisWindow to utilize the Mapper's bounds instead of the Mapper's input dataset bounds.